refactor: extract UI→API integration tests into standalone spec with no storageState
