+++ frontend/e2e/food-catalog.spec.ts
@@ -272,72 +272,3 @@ test.describe('Food Catalog — API Contract & Enum Validation', () => {
     expect(body.data).toHaveProperty('fat');
   });
 });
-
-test.describe('Food Catalog — UI→API Integration', () => {
-  let accessToken: string;
-
-  test.beforeEach(async ({ page, request }) => {
-    const email = uniqueEmail();
-    const password = 'SenhaSegura123!';
-    const result = await signupViaApi(request, email, password);
-    accessToken = result.accessToken;
-    await completeOnboardingViaApi(request, accessToken);
-
-    await page.goto('/login');
-    await page.waitForLoadState('networkidle');
-    await page.evaluate((token) => {
-      localStorage.setItem(
-        'nutriai-auth',
-        JSON.stringify({
-          state: {
-            isAuthenticated: true,
-            accessToken: token,
-            user: {
-              id: 'e2e-int',
-              name: 'Dra. Int',
-              email: 'int@test.com',
-              role: 'NUTRITIONIST',
-              onboardingCompleted: true,
-            },
-          },
-          version: 0,
-        }),
-      );
-    }, accessToken);
-    await page.goto('/foods');
-    await page.waitForLoadState('networkidle');
-  });
-
-  test('E2E-FC-16: Create food via UI sends correct enum keys to API', async ({
-    page,
-    request,
-  }) => {
-    await page.getByRole('button', { name: /novo alimento/i }).click();
-    const modal = page.locator('[role="dialog"], .modal');
-    await expect(modal).toBeVisible({ timeout: 3_000 });
-
-    const nameInput = page.locator('input[placeholder*="Frango"]');
-    await nameInput.fill('Whey Protein E2E');
-
-    const selects = page.locator('select');
-    const categorySelect = selects.first();
-    await categorySelect.selectOption({ label: 'Proteína' });
-
-    const refInput = page.locator('#create-food-ref');
-    await refInput.fill('100');
-
-    const saveBtn = page.getByRole('button', { name: /salvar/i });
-    await saveBtn.click();
-    await expect(modal).not.toBeVisible({ timeout: 5_000 });
-
-    const response = await request.get(`${API_BASE}/foods`, {
-      headers: { Authorization: `Bearer ${accessToken}` },
-      params: { search: 'Whey Protein E2E' },
-    });
-    expect(response.status()).toBe(200);
-    const body = await response.json();
-    const found = body.data?.content?.find((p: { name: string }) => p.name === 'Whey Protein E2E');
-    expect(found).toBeDefined();
-    expect(found.category).toBe('PROTEINA');
-  });
-});
